src/cihai-cli/pyproject.toml
[tool.black]
skip-string-normalization = true

[tool.poetry]
name = "cihai-cli"
version = "0.8.1"
description = "Command line frontend for the cihai CJK language library"
license = "MIT"
authors = ["Tony Narlock <tony@git-pull.com>"]
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: MIT License",
  "Environment :: Web Environment",
  "Intended Audience :: Developers",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: Implementation :: PyPy",
  "Topic :: Software Development :: Internationalization",
  "Topic :: System :: Shells",
  "Topic :: Utilities"
]

packages = [
    { include = "cihai_cli" }
]

homepage = "https://cihai-cli.git-pull.com"
readme = "README.md"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/cihai/cihai-cli/issues"
Documentation = "https://cihai-cli.git-pull.com"
Repository = "https://github.com/cihai/cihai-cli"

[tool.poetry.dependencies]
python = "^3.6"
click = ">=7<8"
cihai = "~=0.12.0"
PyYAML = "<6,>=3.12"

[tool.poetry.dev-dependencies]
### Docs ###
sphinx = "*"
alagitpull = "^0.1.0"
sphinx-issues = "^1.2.0"
sphinx-click = "*"
sphinx-autodoc-typehints = "^1.12.0"
myst_parser = "^0.15.0"

### Testing ###
pytest = "*"
pytest-rerunfailures = "*"

### Coverage ###
codecov = "*"
coverage = "*"
pytest-cov = "*"

### Format ###
black = {version="==21.6b0", python="^3.6.2"}
isort = "*"

### Lint ###
flake8 = "*"

[tool.poetry.scripts]
cihai = 'cihai_cli.cli:cli'

[tool.poetry.extras]
docs = ["sphinx", "myst_parser", "sphinx-issues", "sphinx-click", "alagitpull", "sphinx-autodoc-typehints"]
test = ["pytest", "pytest-rerunfailures"]
coverage = ["codecov", "coverage", "pytest-cov"]
format = ["black", "isort"]
lint = ["flake8"]
